+++ src/arch/x86/kernel/boot_sys.c
@@ -561,7 +561,7 @@ try_boot_sys(
             modules[i].start,
             modules[i].end,
             modules[i].end - modules[i].start,
-            (char *) modules[i].name
+            (char *) (long)modules[i].name
         );
         if ((int32_t)(modules[i].end - modules[i].start) <= 0) {
             printf("Invalid boot module size! Possible cause: boot module file not found by QEMU\n");
